- resolve #157: Refactoring: extract deployment related functionality of RollbarClient into RollbarDeployClient
